Overall improvements and component modularization.
