+++ package.json
@@ -19,7 +19,7 @@
   "devDependencies": {
     "@tsconfig/node22": "^22.0.1",
     "@types/axios": "^0.9.36",
-    "@types/node": "^22.13.14",
+    "@types/node": "^22.14.0",
     "@vitejs/plugin-vue": "^5.2.3",
     "@vue/eslint-config-prettier": "^10.2.0",
     "@vue/eslint-config-typescript": "^14.5.0",
