+++ tsconfig.json
@@ -1,4 +1,5 @@
 {
+  "types": ["node"],
   "files": [],
   "references": [
     {
